feat(post): add update and delete integration testing

+++ internal/use_case/post_use_case.go
@@ -114,7 +114,7 @@ func (p PostUseCase) Update(request *model_controller.UpdatePostRequest) error {
 	post := &entity.Post{
 		Description: request.Description,
 		ImageUrl:    request.ImageUrl,
-		UpdatedAt:   null.NewTime(time.Now(), true),
+		UpdatedAt:   null.NewTime(time.Now().UTC(), true),
 	}
 
 	if err = p.PostRepository.Update(tx, post, request.ID); err != nil {
@@ -133,6 +133,16 @@ func (p PostUseCase) Delete(request *model_controller.DeletePostRequest) error {
 
 	if err != nil {
 		return errors.New(http.StatusText(http.StatusInternalServerError))
+	}
+
+	total, err := p.PostRepository.CountByID(tx, request.ID)
+
+	if err != nil {
+		return errors.New(http.StatusText(http.StatusInternalServerError))
+	}
+
+	if total < 0 {
+		return errors.New(http.StatusText(http.StatusNotFound))
 	}
 
 	if err = p.PostRepository.Delete(tx, request.ID); err != nil {
